Modify imports for sleap-io Video
