+++ sleap/gui/app.py
@@ -78,7 +78,7 @@ from sleap.gui.widgets.video import QtVideoPlayer
 from sleap.info.summary import StatisticSeries
 from sleap_io.model.instance import Instance
 from sleap.io.dataset import Labels
-from sleap.io.video import available_video_exts
+from sleap.sleap_io_adaptors.video_utils import available_video_exts
 from sleap.prefs import prefs
 from sleap_io.model.skeleton import Skeleton
 from sleap.util import parse_uri_path, get_config_file
