+++ sleap/gui/dataviews.py
@@ -27,7 +27,7 @@ from sleap.gui.commands import CommandContext
 from sleap.gui.state import GuiState
 from sleap.instance import LabeledFrame
 from sleap_io.model.skeleton import Skeleton
-from sleap.io.video import Video
+from sleap_io import Video
 from sleap.sleap_io_adaptors.skeleton_utils import get_symmetry_node
 from sleap.sleap_io_adaptors.instance_utils import get_nodes_from_instance, node_points
