+++ sleap/gui/overlays/tracks.py
@@ -9,7 +9,7 @@ from sleap.gui.overlays.base import BaseOverlay
 from sleap.gui.widgets.video import QtTextWithBackground
 from sleap.instance import LabeledFrame
 from sleap_io.model.instance import Track
-from sleap.io.video import Video
+from sleap_io import Video
 from sleap.prefs import prefs
