+++ sleap/info/write_tracking_h5.py
@@ -34,7 +34,7 @@ import pandas as pd
 from typing import Any, Dict, List, Tuple
 
 from sleap.io.dataset import Labels
-from sleap.io.video import Video
+from sleap_io import Video
 from sleap import PredictedInstance
