+++ sleap/io/legacy.py
@@ -10,7 +10,7 @@ import pandas as pd
 from typing import List
 
 from sleap.util import json_loads
-from sleap.io.video import Video
+from sleap_io import Video
 
 from sleap.instance import (
     LabeledFrame,
