+++ tests/gui/test_commands.py
@@ -34,7 +34,7 @@ from sleap.io.dataset import Labels
 from sleap.io.format.adaptor import Adaptor
 from sleap.io.format.ndx_pose import NDXPoseAdaptor
 from sleap.io.pathutils import fix_path_separator
-from sleap.io.video import Video
+from sleap_io import Video
 from sleap.util import get_package_file
 
 # These imports cause trouble when running `pytest.main()` from within the file
