+++ tests/io/test_convert.py
@@ -1,6 +1,6 @@
 from sleap.io.convert import default_analysis_filename, main as sleap_convert
 from sleap.io.dataset import Labels
-from sleap.io.video import Video
+from sleap_io import Video
 from sleap_io.model.instance import Instance
 from pathlib import PurePath, Path
 import re
