+++ tests/io/test_dataset.py
@@ -8,7 +8,8 @@ import sleap
 from sleap.info.write_tracking_h5 import get_nodes_as_np_strings
 from sleap_io import Skeleton
 from sleap.instance import LabeledFrame
-from sleap.io.video import Video, MediaVideo
+from sleap_io import Video
+from sleap_io.io.video_reading import MediaVideo
 from sleap.io.dataset import Labels, load_file
 from sleap.io.format.ndx_pose import NDXPoseAdaptor
 from sleap.io.format import filehandle
