+++ tests/io/test_formats.py
@@ -7,7 +7,7 @@ from numpy.testing import assert_array_equal
 import pytest
 import nixio
 
-from sleap.io.video import Video
+from sleap_io import Video
 from sleap.instance import LabeledFrame, PredictedInstance
 from sleap.io.dataset import Labels
 from sleap.io.format import read, dispatch, adaptor, text, genericjson, hdf5, filehandle
